get recipients
