+++ app/pages/sections/catalogue-solutions/select/recipient/controller.test.js
@@ -1,9 +1,8 @@
 import { getData } from 'buying-catalogue-library';
-import { orderApiUrl, solutionsApiUrl } from '../../../../../config';
+import { solutionsApiUrl } from '../../../../../config';
 import { logger } from '../../../../../logger';
 import {
   getRecipientPageContext,
-  getRecipients,
   getSolution,
   validateRecipientForm,
   getServiceRecipientName,
@@ -26,25 +25,6 @@ describe('catalogue-solutions select-solution controller', () => {
 
       expect(contextCreator.getContext.mock.calls.length).toEqual(1);
       expect(contextCreator.getContext).toHaveBeenCalledWith({ orderId: 'order-1', solutionName: 'Solution One' });
-    });
-  });
-
-  describe('getRecipients', () => {
-    afterEach(() => {
-      getData.mockReset();
-    });
-
-    it('should call getData once with the correct params', async () => {
-      getData
-        .mockResolvedValueOnce({ data: {} });
-
-      await getRecipients({ orderId: 'order-1', accessToken: 'access_token' });
-      expect(getData.mock.calls.length).toEqual(1);
-      expect(getData).toHaveBeenCalledWith({
-        endpoint: `${orderApiUrl}/api/v1/orders/order-1/sections/service-recipients`,
-        accessToken: 'access_token',
-        logger,
-      });
     });
   });
